feat: useHeader
* useHeader 커스텀 훅을 이용해 헤더 컴포넌트로부터 액션, 상태 로직 분리
