+++ frontend/src/common/components/portal/index.jsx
@@ -1,7 +1,6 @@
 import ReactDOM from "react-dom";
 
 const Portal = ({ elementId, children }) => {
-  console.log(elementId);
   const potal = document.getElementById(elementId);
   return potal && ReactDOM.createPortal(children, potal);
 };
